Moving input/event-related code to separate source
